Use `fs` to write to stdout
`process.stdout.write` writes to a Node.js stream, so it does a bunch of
work before anything is actually written. This patch uses `fs.writeSync`
to write to the stdout fd.

If you write a lot of stuff very fast `writeSync` will start throwing
EAGAIN errors. There might be a better way to wait for the file to be
writable again, but this patch just catches those errors and ignores
them.

This speeds it up by about 7x for me, from just below 500KiB/s to
3.5~4MiB/s.